Refactor: Rename DoxygenXmlParser to Driver
